fix(gateway): رفع خطای compilation در downloadFile
باگ: source.read(buffer) مقدار Int برمی‌گرداند ولی با -1L (Long) مقایسه شده بود
تعمیر: تغییر -1L به -1 و حذف .toInt() اضافی

خطا: Operator '==' cannot be applied to 'kotlin.Int' and 'kotlin.Long'

Build حالا با موفقیت compile می‌شود.

+++ app/src/main/java/com/hermes/android/gateway/OkHttpGatewayClient.kt
@@ -387,8 +387,8 @@ class OkHttpGatewayClient @Inject constructor(
                         var bytesRead: Long = 0
                         while (true) {
                             val read = source.read(buffer)
-                            if (read == -1L) break
-                            output.write(buffer, 0, read.toInt())
+                            if (read == -1) break
+                            output.write(buffer, 0, read)
                             bytesRead += read
                         }
                         Timber.d("[Gateway] Downloaded $bytesRead bytes to ${tempFile.absolutePath}")
